+++ src/main/java/programacion3/parcial2/parcial_app/ejercicio1/model/Estudiante.java
@@ -6,11 +6,11 @@ public class Estudiante implements Serializable {
     private static final long serialVersionUID = 1L;
     private String codigo;
     private String nombre;
-    private int nota1;
-    private int nota2;
-    private int nota3;
+    private double nota1;
+    private double nota2;
+    private double nota3;
 
-    public Estudiante(String codigo, String nombre, int nota1, int nota2, int nota3) {
+    public Estudiante(String codigo, String nombre, double nota1, double nota2, double nota3) {
         this.codigo = codigo;
         this.nombre = nombre;
         this.nota1 = nota1;
@@ -38,7 +38,7 @@ public class Estudiante implements Serializable {
         this.nombre = nombre;
     }
 
-    public int getNota1() {
+    public double getNota1() {
         return nota1;
     }
 
@@ -46,7 +46,7 @@ public class Estudiante implements Serializable {
         this.nota1 = nota1;
     }
 
-    public int getNota2() {
+    public double getNota2() {
         return nota2;
     }
 
@@ -54,7 +54,7 @@ public class Estudiante implements Serializable {
         this.nota2 = nota2;
     }
 
-    public int getNota3() {
+    public double getNota3() {
         return nota3;
     }
